Fix creature loot option when Wowhead does not have count info
Examples are creatures that have a ?? count with a ?? percentage. It will now put that chance to 0 with an warning so the user understands they have to change it manually.

Not the cleanest, but it was better than generating half an SQL file with no reason as to why it failed.

Better fix is most likely to allow user to choose if missing Wowhead loot data is outputted to SQL or not.